Surgically revert 41a355e3210c389f2ae1f198333e1f1f86973795 (Modify vTable structure: store concrete object's methods in the vTable)

+++ include/wisey/IConcreteObjectType.hpp
@@ -89,7 +89,7 @@ public:
   /**
    * Returns the interface index in the list of flattened interface hierarchy
    */
-  static unsigned long getInterfaceIndex(IConcreteObjectType* object, Interface* interface);
+  static int getInterfaceIndex(IConcreteObjectType* object, Interface* interface);
   
   /**
    * Initialize vTable of a newly created concrete object
